requirements.txt
pandas
numpy
tqdm
pytorch
datasets
transformers